fix: afcen_score Decimal serializes as string — change to float in schema
Decimal fields serialize to JSON strings in Pydantic; score fields must be
float so the frontend receives numbers and .toFixed() works correctly.
Also wrap score in Number() cast as a defensive guard.

+++ backend/app/schemas/pipeline_schemas.py
@@ -105,11 +105,11 @@ class ProjectPipelineRead(BaseModel):
     investment_size: Decimal
     currency: str = "USD"
     
-    # Scores
+    # Scores — must be float so JSON serializes as numbers, not strings
     readiness_score: float = 0.0
-    afcen_score: Optional[Decimal] = None
-    strategic_alignment_score: Optional[Decimal] = None
-    regional_impact_score: Optional[Decimal] = None
+    afcen_score: Optional[float] = None
+    strategic_alignment_score: Optional[float] = None
+    regional_impact_score: Optional[float] = None
     
     # Metadata
     pillar: Optional[str] = None
